+++ src/Concerto/PanelBundle/Resources/starter_content/src/template/footer/footer.js
@@ -1,0 +1,3 @@
+testRunner.controllerProvider.register("footer", function($scope) {
+  $scope.footerContent = testRunner.R.footerContent ? testRunner.R.footerContent : 'Created with <a href="http://www.concertoplatform.com" target="_blank">Concerto Platform</a>';
+});