+++ src/Concerto/PanelBundle/Resources/starter_content/src/template/header/header.js
@@ -1,0 +1,4 @@
+testRunner.controllerProvider.register("header", function($scope) {
+  $scope.title = testRunner.R.title;
+  $scope.logo = testRunner.R.logo;
+});