+++ src/Concerto/PanelBundle/Command/ConcertoGitPullCommand.php
@@ -4,6 +4,7 @@ namespace Concerto\PanelBundle\Command;
 
 use Concerto\PanelBundle\Service\GitService;
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Process\Process;
@@ -15,6 +16,8 @@ class ConcertoGitPullCommand extends Command
 
     /** @var OutputInterface */
     private $output;
+    private $username;
+    private $email;
 
     public function __construct(GitService $gitService)
     {
@@ -25,12 +28,16 @@ class ConcertoGitPullCommand extends Command
     protected function configure()
     {
         $this->setName("concerto:git:pull")->setDescription("Performs git pull");
+        $this->addArgument("username", InputArgument::REQUIRED, "Commit username");
+        $this->addArgument("email", InputArgument::REQUIRED, "Commit email");
     }
 
     private function getPullCommand()
     {
         $exec = $this->gitService->getGitExecPath();
-        return "$exec pull";
+        $usernameCmd = "$exec config user.name \"" . $this->username . "\"";
+        $emailCmd = "$exec config user.email \"" . $this->email . "\"";
+        return "$usernameCmd; $emailCmd; $exec pull --commit --no-edit";
     }
 
     private function pull()
@@ -49,6 +56,8 @@ class ConcertoGitPullCommand extends Command
     {
         $this->output = $output;
         $this->localGitRepoPath = $this->gitService->getGitRepoPath();
+        $this->username = $input->getArgument("username");
+        $this->email = $input->getArgument("email");
 
         chdir($this->localGitRepoPath);