+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 * **concerto.template.show** now accepts **cookies** input parameter (name-value pair list)
 * added **CONCERTO_BEHIND_PROXY** env and **behind_proxy** config option for SAML toolkit
 
+#### Starter content
+
+* added **instructions** block to **test** and **form** template
+
 #### Fixes
 
 * Disabled escape characters for both CSV import and export